Unity 2023

+++ Library/Bee/artifacts/WebGL/il2cppOutput/cpp/System__1.cpp
@@ -21747,7 +21747,7 @@ IL_0015:
 		int16_t L_8 = V_2;
 		NullCheck(L_7);
 		StringBuilder_t* L_9;
-		L_9 = StringBuilder_Append_m71228B30F05724CD2CD96D9611DCD61BFB96A6E1(L_7, ((int32_t)(uint16_t)((-L_8))), NULL);
+		L_9 = StringBuilder_Append_m71228B30F05724CD2CD96D9611DCD61BFB96A6E1(L_7, ((int32_t)(uint16_t)((-((int32_t)L_8)))), NULL);
 		int32_t L_10 = V_1;
 		V_1 = ((int32_t)il2cpp_codegen_add(L_10, 1));
 	}
